chore(OccurrenceSpecies): update model, factory, migration
